requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
dataclasses-json==0.6.7
deepgram-sdk==4.1.0
deprecation==2.1.0
exceptiongroup==1.3.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
packaging==25.0
propcache==0.3.1
python-dotenv==1.1.0
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.13.2
websockets==15.0.1
yarl==1.20.0
yt-dlp==2025.5.22
